Change handling of variables in register types
Disallow free variables, but allow global constants and variables in
indices of field declarations.  Note that global variables in field
indices will not be evaluated at register declaration time, but every
time the field is accessed (the type checker desugars register field
accesses to slice expressions, copying the indices).

This can be used, for example, to have the location of a named field
depend on the current exception level:

    __register 32 { (if EL == 0 then 0 else 1)+:1 foo } bar;

Probably a cleaner approach would be to forbid all uses of variables in
register types and implement fields with dynamic positions using
getter/setter functions, but this would require changes to some existing
ASL code.